requirements.txt
streamlit
pymupdf
chromadb
sentence-transformers
easyocr
pillow
numpy
requests